+++ src/components/Form/Form.context.tsx
@@ -8,9 +8,12 @@ import {
   isUndefined,
   type DeepKeyOf,
   type Logger,
+  type Nullable,
   type Nullish,
+  type NullishPrimitives,
   type VoidFn,
 } from "@ubloimmo/front-util";
+import { merge } from "lodash";
 import {
   FormEventHandler,
   createContext,
@@ -26,6 +29,8 @@ import {
 
 import {
   buildFormText,
+  isFormCustomContent,
+  isFormCustomField,
   isFormDivider,
   isFormText,
   isSchemaFieldRequired,
@@ -59,8 +64,20 @@ import type {
   FormContent,
   BuiltFormContent,
   ComputeFormValidationFn,
+  BuildCustomFieldPropsFn,
+  FormCustomFieldProps,
+  BuiltFormCustomFieldProps,
+  CustomFormInputProps,
+  FormFieldLayout,
+  BuiltFormFieldLayoutProps,
+  FormDefaultProps,
+  DefaultFormLayoutProps,
+  BuildFormFieldLayoutFn,
+  UseFormLayoutReturn,
+  FormFieldLayoutHiddenFn,
 } from "./Form.types";
 import type { InputOnChangeFn, InputType } from "../Input";
+import type { GridEndPosition } from "@layouts";
 
 /**
  * Custom form hook
@@ -100,32 +117,41 @@ const useFormData = <TData extends object>(
    * Loads form data from query into `data` and `initialData` states
    * white updating `isLoading` state
    */
-  const loadFormData = useCallback(async () => {
-    // TODO: Merge with default values
-    if (!isFunction<FormQueryFn<TData>>(props.query)) {
-      if (isObject(props.query)) {
-        setData(props.query);
-        setInitialData(props.query);
+  const loadFormData = useCallback(
+    async (mergeWithCurrent?: boolean) => {
+      if (!isFunction<FormQueryFn<TData>>(props.query)) {
+        if (isObject(props.query)) {
+          const newData = mergeWithCurrent
+            ? merge({ ...initialData }, { ...props.query })
+            : { ...props.query };
+          setData(newData);
+          setInitialData(newData);
+        }
+        return;
       }
-      return;
-    }
-    setIsLoading(true);
-    try {
-      const data: FormData<TData> = await props.query();
-      setData(data);
-      setInitialData(data);
-    } catch (e) {
-      logger.error(e);
-      logger.warn("Failed to load form data");
-    }
-    setIsLoading(false);
-  }, [props, logger]);
+      setIsLoading(true);
+      try {
+        const data: FormData<TData> = await props.query();
+        const newData = mergeWithCurrent
+          ? merge({ ...initialData }, data)
+          : data;
+        setData(newData);
+        setInitialData(newData);
+      } catch (e) {
+        logger.error(e);
+        logger.warn("Failed to load form data");
+      }
+      setIsLoading(false);
+    },
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    [props.query, logger]
+  );
 
   /**
    * Effect used for loading initial form data if query is a function
    */
   useEffect(() => {
-    loadFormData();
+    loadFormData(true);
   }, [props.query, loadFormData]);
 
   /**
@@ -247,13 +273,55 @@ const useFormValidation = <TData extends object>(
     if (modifiers.validateOnChange) {
       triggerFormValidation();
     }
-  }, [data, modifiers, triggerFormValidation]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [data, triggerFormValidation]);
 
   return {
     ...formValidation,
     schema,
     triggerFormValidation,
     computeFormValidation,
+  };
+};
+
+const useFormLayout = (
+  formLayout: DefaultFormLayoutProps
+): UseFormLayoutReturn => {
+  const columns = useMemo(() => {
+    // only allow even number of columns
+    const minColumns = Math.max(formLayout.columns, 2);
+    const evenColumns = Math.ceil(minColumns / 2) * 2;
+    return evenColumns;
+  }, [formLayout]);
+
+  const buildFormFieldLayout = useCallback<BuildFormFieldLayoutFn>(
+    (fieldLayout?: FormFieldLayout): BuiltFormFieldLayoutProps["layout"] => {
+      const defaultSize = Math.max(1, Math.round(columns / 2));
+
+      const hidden = isFunction<FormFieldLayoutHiddenFn>(fieldLayout?.hidden)
+        ? fieldLayout.hidden()
+        : isBoolean(fieldLayout?.hidden)
+        ? fieldLayout.hidden
+        : false;
+
+      const size = fieldLayout?.size ?? defaultSize;
+
+      const columnEnd: GridEndPosition = `span ${size}`;
+
+      return {
+        ...fieldLayout,
+        hidden,
+        size,
+        columnEnd,
+        readonly: fieldLayout?.readonly ?? false,
+      };
+    },
+    [columns]
+  );
+
+  return {
+    columns,
+    buildFormFieldLayout,
   };
 };
 
@@ -267,12 +335,13 @@ const useFormValidation = <TData extends object>(
  * @param {UseFormValidationReturn<TData>} validation - The form validation.
  * @param {FormModifers} modifiers - The form modifiers.
  * @param {FormContent<TData, TType>[]} content - The form content.
- * @return {BuiltFormContent<InputType>[]} - The built field props for each form field.
+ * @return {BuiltFormContent<TData, InputType>[]} - The built field props for each form field.
  */
 const useFormFields = <TData extends object>(
   formData: UseFormDataReturn<TData>,
   validation: UseFormValidationReturn<TData>,
   modifiers: FormModifers,
+  formLayout: UseFormLayoutReturn,
   content?: FormContent<TData>[]
 ): BuiltFormContent<InputType>[] => {
   /**
@@ -364,8 +433,8 @@ const useFormFields = <TData extends object>(
         type: type as TType,
         disabled: disabled || modifiers.disabled,
         required: isFieldRequired(source, required),
-        layout,
-      } as BuiltFieldProps<TType>;
+        layout: formLayout.buildFormFieldLayout(layout),
+      };
     },
     [
       getFieldErrorProps,
@@ -373,6 +442,50 @@ const useFormFields = <TData extends object>(
       getFieldValue,
       modifiers.disabled,
       isFieldRequired,
+      formLayout,
+    ]
+  );
+
+  const buildCustomFieldProps = useCallback<BuildCustomFieldPropsFn<TData>>(
+    (
+      formCustomField: FormCustomFieldProps<TData>
+    ): BuiltFormCustomFieldProps => {
+      const {
+        source,
+        CustomInput,
+        kind: _kind,
+        layout,
+        onChange,
+        disabled,
+        errorText,
+        error,
+        required,
+        ...rest
+      } = formCustomField;
+
+      return {
+        ...rest,
+        ...getFieldErrorProps(source, error, errorText),
+        CustomInput,
+        onChange: propagateChange(
+          source,
+          onChange as Nullable<InputOnChangeFn<InputType>>
+        ) as CustomFormInputProps<NullishPrimitives>["onChange"],
+        value: getFieldValue<DeepKeyOf<FormData<TData>>>(
+          source as DeepKeyOf<FormData<TData>>
+        ),
+        disabled: disabled || modifiers.disabled,
+        required: isFieldRequired(source, required),
+        layout: formLayout.buildFormFieldLayout(layout),
+      };
+    },
+    [
+      formLayout,
+      getFieldErrorProps,
+      getFieldValue,
+      isFieldRequired,
+      modifiers.disabled,
+      propagateChange,
     ]
   );
 
@@ -383,11 +496,13 @@ const useFormFields = <TData extends object>(
   return useMemo<BuiltFormContent<InputType>[]>(() => {
     if (!content || !content.length) return [];
     return content.map((content) => {
-      if (isFormDivider(content)) return content;
+      if (isFormDivider(content) || isFormCustomContent(content))
+        return content;
       if (isFormText(content)) return buildFormText(content);
+      if (isFormCustomField(content)) return buildCustomFieldProps(content);
       return buildFieldProps(content);
     });
-  }, [buildFieldProps, content]);
+  }, [buildCustomFieldProps, buildFieldProps, content]);
 };
 
 /**
@@ -546,7 +661,7 @@ const useFormEditState = (modifiers: FormModifers): UseFormEditStateReturn => {
  * @see {@link useFormData}, {@link useFormValidation}, {@link useFormEditState} {@link useFormModifiers}, {@link useFormFields}, {@link useFormSubmission}
  */
 export const useForm = <TData extends object>(
-  props: FormProps<TData>,
+  { columns, modal, ...props }: FormDefaultProps<TData>,
   logger: Logger
 ): FormContext<TData> => {
   const formData = useFormData<TData>(props, logger);
@@ -557,10 +672,12 @@ export const useForm = <TData extends object>(
     formModifiers
   );
   const formEditState = useFormEditState(formModifiers);
+  const formLayout = useFormLayout({ columns, modal });
   const content = useFormFields(
     formData,
     formValidation,
     formModifiers,
+    formLayout,
     props.content
   );
   const formSubmission = useFormSubmission(
@@ -578,6 +695,7 @@ export const useForm = <TData extends object>(
     ...formSubmission,
     ...formEditState,
     ...formModifiers,
+    ...formLayout,
     content,
   };
 };
@@ -601,6 +719,13 @@ const defaultFormContext: FormContext<object> = {
   ...defaultFormValidation,
   triggerFormValidation: () => {},
   computeFormValidation: () => defaultFormValidation,
+  columns: 2,
+  buildFormFieldLayout: () => ({
+    readonly: false,
+    hidden: false,
+    size: 1,
+    columnEnd: "span 1",
+  }),
 };
 
 const InternalFormContext =
@@ -621,7 +746,7 @@ export const useFormContext = <TData extends object>(): FormContext<TData> => {
 };
 
 export const FormProvider = <TData extends object>(
-  props: FormProps<TData> & { children: ReactNode }
+  props: FormDefaultProps<TData> & { children: ReactNode }
 ): JSX.Element => {
   const logger = useLogger("Form Context");
   const context = useForm<TData>(props, logger);