+++ src/components/Form/components/FormFieldRenderer.component.tsx
@@ -1,32 +1,59 @@
-import { isString, type Nullable } from "@ubloimmo/front-util";
-import { useMemo } from "react";
+import { isFunction, isString, type Nullable } from "@ubloimmo/front-util";
+import { FC, useMemo } from "react";
 import styled from "styled-components";
 
+import { FormCustomContent } from "./FormCustomContent.component";
+import { FormCustomField } from "./FormCustomField.component";
 import { FormDivider } from "./FormDivider.component";
 import { FormField } from "./FormField.component";
 import { FormText } from "./FormText.component";
 import { useFormContext } from "../Form.context";
 import { formFieldListContainerStyles } from "../Form.styles";
 import {
+  isBuiltCustomFormField,
   isBuiltFormField,
   isBuiltFormText,
+  isFormCustomContent,
   isFormDivider,
 } from "../Form.utils";
 
 import { GridLayout } from "@layouts";
 
+import type { BuiltFormCustomContentProps } from "../Form.types";
 import type { DividerProps } from "@/components/Divider";
 
 export const FormFieldRenderer = <TData extends object>() => {
-  const { content } = useFormContext<TData>();
+  const { content, columns } = useFormContext<TData>();
 
   const renderedContent = useMemo<Nullable<JSX.Element>[]>(() => {
     return content.map((contentItem, index) => {
       if (isBuiltFormField(contentItem)) {
         return <FormField {...contentItem} key={`form-field-${index}`} />;
       }
+      if (isBuiltCustomFormField(contentItem)) {
+        return (
+          <FormCustomField
+            {...contentItem}
+            key={`form-custom-field-${index}`}
+          />
+        );
+      }
       if (isBuiltFormText(contentItem)) {
-        return <FormText {...contentItem} key={`form-field-${index}`} />;
+        return <FormText {...contentItem} key={`form-text-${index}`} />;
+      }
+      if (isFormCustomContent(contentItem)) {
+        const contentProps: BuiltFormCustomContentProps = {
+          content: isFunction<FC>(contentItem)
+            ? contentItem
+            : contentItem.content,
+          kind: "content",
+        };
+        return (
+          <FormCustomContent
+            {...contentProps}
+            key={`form-custom-content-${index}`}
+          />
+        );
       }
       if (!isFormDivider(contentItem)) return null;
       const dividerProps: DividerProps = isString(contentItem)
@@ -38,7 +65,7 @@ export const FormFieldRenderer = <TData extends object>() => {
 
   return (
     <FieldListContainer
-      columns={2}
+      columns={columns}
       gap="s-6"
       testId="form-field-list"
       overrideTestId